recipe::cookTime: change to int

+++ Repository/Models/Recipes/CreateRecipe.cs
@@ -11,7 +11,7 @@ namespace Repository.Models.Recipes
         public string? Description { get; set; }
         public int? ServingSize { get; set; }
         public string? PhotoUrl { get; set; }
-        public decimal? CookTime { get; set; }
+        public int? CookTime { get; set; }
         public IEnumerable<CreateRecipeMaterial> Ingredients { get; set; } = new List<CreateRecipeMaterial>();
         public IEnumerable<CreateRecipeMaterial> Equipment { get; set; } = new List<CreateRecipeMaterial>();
         public IEnumerable<CreateRecipeStep> RecipeSteps { get; set; } = new List<CreateRecipeStep>();
